# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Q-Vuln Post-Quantum Security Scanner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /github/scan:
    post:
      operationId: scanGithubRepo
      tags: [github]
      summary: Scan a public GitHub repository for quantum vulnerabilities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubScanBody"
      responses:
        "200":
          description: GitHub scan results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubScanResult"
        "400":
          description: Invalid URL
        "404":
          description: Repository not found
        "429":
          description: GitHub API rate limit
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        language:
          type: string
        riskScore:
          type: integer
          nullable: true
        lastScanAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        totalScans:
          type: integer
        criticalCount:
          type: integer
        alertCount:
          type: integer
        cleanCount:
          type: integer
      required:
        - id
        - name
        - language
        - createdAt
        - totalScans
        - criticalCount
        - alertCount
        - cleanCount
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        language:
          type: string
        code:
          type: string
      required:
        - name
        - language
        - code
...
    Finding:
      type: object
      properties:
        id:
          type: integer
        scanId:
          type: integer
        fileName:
          type: string
        lineNumber:
          type: integer
        severity:
          type: string
          enum: [critical, alert, safe]
        algorithm:
          type: string
        codeSnippet:
          type: string
        nistReplacement:
          type: string
          nullable: true
        nistStandard:
          type: string
          nullable: true
        effortHours:
          type: number
        explanation:
          type: string
          nullable: true
      required:
        - id
        - scanId
        - fileName
        - lineNumber
        - severity
        - algorithm
        - codeSnippet
        - effortHours
...
    GithubScanBody:
      type: object
      properties:
        repoUrl:
          type: string
          description: Public GitHub repository URL
      required:
        - repoUrl
    GithubFileResult:
      type: object
      properties:
        path:
          type: string
        language:
          type: string
        lines:
          type: integer
        criticalCount:
          type: integer
        alertCount:
          type: integer
      required:
        - path
        - language
        - lines
        - criticalCount
        - alertCount
    GithubScanResult:
      type: object
      properties:
        repoUrl:
          type: string
        owner:
          type: string
        repo:
          type: string
        totalFiles:
          type: integer
        findings:
          type: array
          items:
            $ref: "#/components/schemas/Finding"
        criticalCount:
          type: integer
        alertCount:
          type: integer
        cleanCount:
          type: integer
        riskScore:
          type: integer
        totalLines:
          type: integer
        totalEffortHours:
          type: number
          nullable: true
        executiveSummary:
          type: string
        fileResults:
          type: array
          items:
            $ref: "#/components/schemas/GithubFileResult"
      required:
        - repoUrl
        - owner
        - repo
        - totalFiles
        - findings
        - criticalCount
        - alertCount
        - cleanCount
        - riskScore
        - totalLines
        - executiveSummary
        - fileResults